# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vibescape API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/posts/{post_id}/comments": {
      "post": {
        "tags": [
          "comments"
        ],
        "summary": "Create Comment",
        "description": "Comment on a post, or reply to a comment by sending `parent_id`.",
        "operationId": "create_comment_api_v1_posts__post_id__comments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "List Comments",
        "description": "Top-level comments on a post, oldest first.\n\nReplies are not interleaved here \u2014 each carries a `replies_count`, and the\nthread is fetched from `/comments/{id}/replies`. Mixing them in would let one\nbusy thread push every other comment off the first page.",
        "operationId": "list_comments_api_v1_posts__post_id__comments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Post Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedComments"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comments/{comment_id}/replies": {
      "post": {
        "tags": [
          "comments"
        ],
        "summary": "Create Reply",
        "description": "Reply to a comment.\n\nThreads stay one level deep: replying to a reply attaches to the same top-level\ncomment rather than nesting further.",
        "operationId": "create_reply_api_v1_comments__comment_id__replies_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Comment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReplyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "List Replies",
        "description": "Replies under one comment, oldest first.",
        "operationId": "list_replies_api_v1_comments__comment_id__replies_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Comment Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedComments"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comments/{comment_id}": {
      "delete": {
        "tags": [
          "comments"
        ],
        "summary": "Delete Comment",
        "description": "Delete a comment.\n\nAllowed for the comment's author, and for the owner of the post it is on so they\ncan moderate their own content. Anyone else gets 403.\n\nDeleting a top-level comment takes its replies with it.",
        "operationId": "delete_comment_api_v1_comments__comment_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Comment Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hashtags/search": {
      "get": {
        "tags": [
          "hashtags"
        ],
        "summary": "Search Hashtags",
        "description": "Type-ahead over tags. Prefix match, most used first.\n\nRegistered before `/hashtags/{tag}` so the literal path wins \u2014 otherwise\n\"search\" would be read as a tag name.",
        "operationId": "search_hashtags_api_v1_hashtags_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HashtagOut"
                  },
                  "title": "Response Search Hashtags Api V1 Hashtags Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hashtags/{tag}": {
      "get": {
        "tags": [
          "hashtags"
        ],
        "summary": "Read Hashtag",
        "description": "One tag and how many posts carry it.",
        "operationId": "read_hashtag_api_v1_hashtags__tag__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashtagOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hashtags/{tag}/posts": {
      "get": {
        "tags": [
          "hashtags"
        ],
        "summary": "List Hashtag Posts",
        "description": "Posts carrying a tag, newest first, as a grid.\n\nReturns an empty page rather than 404 for a tag nobody has used: a tag landing\npage reached by tapping text in a caption should show \"nothing here\", not an\nerror.",
        "operationId": "list_hashtag_posts_api_v1_hashtags__tag__posts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tag"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGrid"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "post_id": {
            "type": "string",
            "format": "uuid",
            "title": "Post Id"
          },
          "author": {
            "$ref": "#/components/schemas/PostAuthor"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "replies_count": {
            "type": "integer",
            "title": "Replies Count",
            "default": 0
          },
          "mentions": {
            "items": {
              "$ref": "#/components/schemas/PostAuthor"
            },
            "type": "array",
            "title": "Mentions"
          },
          "can_delete": {
            "type": "boolean",
            "title": "Can Delete",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "post_id",
          "author",
          "body",
          "created_at"
        ],
        "title": "CommentOut"
      },
...
      "CreateCommentRequest": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 2200,
            "minLength": 1,
            "title": "Body"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "body"
        ],
        "title": "CreateCommentRequest"
      },
...
      "CreateReplyRequest": {
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 2200,
            "minLength": 1,
            "title": "Body"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "body"
        ],
        "title": "CreateReplyRequest",
        "description": "Body for POST /comments/{id}/replies \u2014 the parent comes from the path."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HashtagOut": {
        "properties": {
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "posts_count": {
            "type": "integer",
            "title": "Posts Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "tag"
        ],
        "title": "HashtagOut"
      },
...
      "PaginatedComments": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommentOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PaginatedComments"
      },
      "PaginatedGrid": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PostGridItem"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PaginatedGrid"
      },
...
      "PostAuthor": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "PostAuthor",
        "description": "Denormalised author, matching what the feed card already renders.\n\n`Posts.ts` reads `username` and `profileImage` straight off each post rather than\njoining, so the payload carries them inline."
      },
      "PostGridItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned",
            "default": false
          },
          "likes_count": {
            "type": "integer",
            "title": "Likes Count",
            "default": 0
          },
          "comments_count": {
            "type": "integer",
            "title": "Comments Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "PostGridItem",
        "description": "Profile-grid projection \u2014 matches `ProfilePosts.ts` (id, image, pinned)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}